requirements.txt
pyyaml
rasterio
scikit-learn
numpy
pandas
matplotlib
scipy
geopandas
shapely
